Requirements

+++ MVC/Controllers/MovieController.cs
@@ -5,25 +5,43 @@ using MVC.Services;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
-using System.Diagnostics;
+using MVC.Interfaces;
 
 namespace MVC.Controllers;
 
 public class MovieController : Controller
 {
-    protected readonly MovieService _movieService;
-    protected readonly UserService _userService;
+    protected readonly IMovieService _movieService;
+    protected readonly IUserService _userService;
+    protected readonly IMovieCreatorService _movieCreatorService;
+    protected readonly IUserMovieService _userMovieService;
+    protected readonly IReportService _reportService;
+    protected readonly IMovieCountryService _movieCountryService;
+    protected readonly ICountryService _countryService;
+    protected readonly IGenreService _genreService;
+    protected readonly IMovieGenreService _movieGenreService;
 
-    public MovieController(MovieService movieService, UserService userService)
+    public MovieController(IMovieService movieService, IUserService userService, 
+                            IMovieCreatorService movieCreatorService, IUserMovieService userMovieService,
+                            IReportService reportService, IMovieCountryService movieCountryService,
+                            ICountryService countryService, IMovieGenreService movieGenreService,
+                            IGenreService genreService)
     {
         _movieService = movieService;
         _userService = userService;
+        _movieCreatorService = movieCreatorService;
+        _userMovieService = userMovieService;
+        _reportService = reportService;
+        _movieCountryService = movieCountryService;
+        _countryService = countryService;
+        _genreService = genreService;
+        _movieGenreService = movieGenreService;
     }
 
     public async Task<IActionResult> Create()
     {
-        ViewBag.Genres = await _movieService.GetGenresDictionaryAsync();
-        ViewBag.Countries = await _movieService.GetCountriesDictionaryAsync();
+        ViewBag.Genres = await _genreService.GetGenresDictionaryAsync();
+        ViewBag.Countries = await _countryService.GetCountriesDictionaryAsync();
         return View();
     }
 
@@ -33,10 +51,12 @@ public class MovieController : Controller
     {
         if (!ModelState.IsValid)
         {
-            ViewBag.Genres = await _movieService.GetGenresDictionaryAsync();
-            ViewBag.Countries = await _movieService.GetCountriesDictionaryAsync();
+            ViewBag.Genres = await _genreService.GetGenresDictionaryAsync();
+            ViewBag.Countries = await _countryService.GetCountriesDictionaryAsync();
             return View(movieViewModel);
         }
+        var currentUser = await _userService.GetCurrentUserAsync(User);
+        if (currentUser == null) return RedirectToAction("Login", "User");
 
         Movie movie = new Movie(
             movieViewModel.Cover,
@@ -47,24 +67,15 @@ public class MovieController : Controller
             movieViewModel.Description
         );
 
-        var result = await _movieService.AddMovieAsync(movie);
-        if (!result.Success)
-        {
-            ViewBag.ErrorMessage = result.ErrorMessage;
-            ViewBag.Genres = await _movieService.GetGenresDictionaryAsync();
-            ViewBag.Countries = await _movieService.GetCountriesDictionaryAsync();
-            return View(movieViewModel);
-        }
-
-        await _movieService.ConnectToGenre(movieViewModel, movie);
-        await _movieService.ConnectToCountry(movieViewModel, movie);
-        var currentUser = await _userService.GetCurrentUserAsync(User);
-        if (currentUser != null) await _movieService.ConnectToCreator(movie, currentUser);
+        await _movieService.AddMovieAsync(movie);
+        await _movieCreatorService.AddMovieCreatorAsync(movie.Id, currentUser.Id);
+        await _movieGenreService.AddMovieGenre(movieViewModel, movie);
+        await _movieCountryService.AddMovieCountry(movieViewModel, movie);
 
         return RedirectToAction("ViewRating", "Movie");
     }
 
-    [Authorize(Roles = "Admin")]
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> Update(int id)
     {
@@ -72,6 +83,12 @@ public class MovieController : Controller
 
         if (movie == null)
             return NotFound();
+
+        var currentUser = await _userService.GetCurrentUserAsync(User);
+        
+        if (!(User.IsInRole("Admin") || 
+            (currentUser != null && await _movieCreatorService.IsCreatorAsync(id, currentUser.Id))))
+            return Forbid();
 
         var movieViewModel = new MovieViewModel
         {
@@ -83,30 +100,36 @@ public class MovieController : Controller
             Description = movie.Description,
             SelectedGenreIds = movie.MovieGenres.Select(mg => mg.GenreId).ToList(),
             SelectedCountryIds = movie.MovieCountries.Select(mc => mc.CountryId).ToList(),
-            Genres = await _movieService.GetGenresDictionaryAsync(),
-            Countries = await _movieService.GetCountriesDictionaryAsync()
+            Genres = await _genreService.GetGenresDictionaryAsync(),
+            Countries = await _countryService.GetCountriesDictionaryAsync()
         };
 
         ViewBag.MovieId = id;
         return View(movieViewModel);
     }
 
-    [Authorize(Roles = "Admin")]
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Update(int id, MovieViewModel movieViewModel)
     {
-        if (!ModelState.IsValid)
-        {
-            movieViewModel.Genres = await _movieService.GetGenresDictionaryAsync();
-            movieViewModel.Countries = await _movieService.GetCountriesDictionaryAsync();
-            ViewBag.MovieId = id;
-            return View(movieViewModel);
-        }
-
         var movie = await _movieService.GetMovieByIdWithRelationsAsync(id);
 
         if (movie == null)
             return NotFound();
+
+        var currentUser = await _userService.GetCurrentUserAsync(User);
+        
+        if (!(User.IsInRole("Admin") || 
+            (currentUser != null && await _movieCreatorService.IsCreatorAsync(id, currentUser.Id))))
+            return Forbid();
+
+        if (!ModelState.IsValid)
+        {
+            movieViewModel.Genres = await _genreService.GetGenresDictionaryAsync();
+            movieViewModel.Countries = await _countryService.GetCountriesDictionaryAsync();
+            ViewBag.MovieId = id;
+            return View(movieViewModel);
+        }
 
         movie.Cover = movieViewModel.Cover;
         movie.Title = movieViewModel.Title;
@@ -115,19 +138,10 @@ public class MovieController : Controller
         movie.Director = movieViewModel.Director;
         movie.Description = movieViewModel.Description;
 
-        _movieService.UpdateMovieGenres(movie, movieViewModel.SelectedGenreIds);
-        _movieService.UpdateMovieCountries(movie, movieViewModel.SelectedCountryIds);
+        _movieGenreService.UpdateMovieGenre(movie, movieViewModel.SelectedGenreIds);
+        _movieCountryService.UpdateMovieCountry(movie, movieViewModel.SelectedCountryIds);
 
-        var result = await _movieService.UpdateMovieAsync(movie);
-        if (!result.Success)
-        {
-            ViewBag.ErrorMessage = result.ErrorMessage;
-
-            movieViewModel.Genres = await _movieService.GetGenresDictionaryAsync();
-            movieViewModel.Countries = await _movieService.GetCountriesDictionaryAsync();
-            ViewBag.MovieId = id;
-            return View(movieViewModel);
-        }
+        await _movieService.UpdateMovieAsync(movie);
 
         return RedirectToAction("ViewRating", "Movie");
     }
@@ -135,7 +149,15 @@ public class MovieController : Controller
     [Authorize(Roles = "Admin")]
     public async Task<IActionResult> Delete(int id)
     {
+        if (!User.IsInRole("Admin")) return Forbid();
+        
         await _movieService.DeleteMovieAsync(id);
+        _movieGenreService.DeleteMovieGenres(id);
+        _movieCountryService.DeleteMovieCountries(id);
+        await _reportService.DeleteReportsForMovieAsync(id);
+        await _userMovieService.DeleteUserMoviesAsync(id);
+        await _movieCreatorService.DeleteMovieCreatorsAsync(id);
+
         return RedirectToAction("ViewRating", "Movie");
     }
 
@@ -153,14 +175,15 @@ public class MovieController : Controller
             var user = await _userService.GetCurrentUserAsync(User);
             if (user != null)
             {
-                var userMovies = (await _userService.GetUserMovies(user, true)).ToList();
-                userMovies.AddRange(await _userService.GetUserMovies(user, false));
+                var userMovies = (await _userMovieService.GetUserMoviesAsync(user.Id, true)).ToList();
+                userMovies.AddRange(await _userMovieService.GetUserMoviesAsync(user.Id, false));
                 
                 ViewBag.IsInUserLists = userMovies.Any(um => um.MovieId == id);
+                ViewBag.IsCreator = await _movieCreatorService.IsCreatorAsync(id, user.Id);
             }
         }
         
-        var reports = await _movieService.GetReportsForMovieAsync(id);
+        var reports = await _reportService.GetReportsForMovieAsync(id);
 
         var reportViewModel = reports.Select(r => new ReportViewModel
         {
@@ -193,5 +216,4 @@ public class MovieController : Controller
         var movies = await _movieService.SearchMoviesAsync(searchTerm);
         return View("ViewRating", movies.OrderByDescending(m => m.Rating).ToList());
     }
-
 }
